Fixed another size_t warning

+++ tests/SimpleStringTest.cpp
@@ -358,7 +358,7 @@ TEST(SimpleString, at)
 TEST(SimpleString, copyInBufferNormal)
 {
 	SimpleString str("Hello World");
-	int bufferSize = str.size()+1;
+	size_t bufferSize = str.size()+1;
 	char* buffer = (char*) malloc(bufferSize);
 	str.copyToBuffer(buffer, bufferSize);
 	STRCMP_EQUAL(str.asCharString(), buffer);
